==> AGN_ver1.0/0.3.split_redshift_bin.ipynb <==
# matpltlib 库以及参数
import matplotlib
import matplotlib.pyplot as plt
from matplotlib.colors import LogNorm
matplotlib.rcParams.update({'font.size': 20})
matplotlib.rcParams.update({'axes.labelsize': 'large'})
matplotlib.rcParams.update({'xtick.labelsize': 'large'})
matplotlib.rcParams.update({'ytick.labelsize': 'large'})
matplotlib.rcParams.update({'legend.fontsize': 'large'})
matplotlib.rcParams.update({'errorbar.capsize': 3})

#其他的库
import numpy as np
import pandas as pd
import scipy as sp 

#封装进度条的函数
from tqdm import tqdm    
#导入自己的函数
import mydef

df=pd.read_csv('./DATA/matched_data_step2.csv')
df

# 按照红移分bin

df.sort_values(by='REDSHIFT',ascending=True)
df['red_bin_index']=pd.cut(df.index,bins=np.arange(0,5214+100,100),right=False,labels=range(53))#RIGHT=FALSE表示左闭右开
df['org_index']=df.index
df_r=df.set_index(['red_bin_index','org_index'])

df_r

# 向其中扩充3+3列方便计算


df_r['g_SDSS-HSC']=df_r['g_SDSS']-df_r['g_HSC']
df_r['r_SDSS-HSC']=df_r['r_SDSS']-df_r['r_HSC']
df_r['i_SDSS-HSC']=df_r['i_SDSS']-df_r['i_HSC']

df_r['MBH']=10**df_r['LOGBH']
df_r['EDD_RATIO']=10**df_r['LOGEDD_RATIO']
df_r['LBOL']=10**df_r['LOGLBOL']

df_r['half_g_SDSS+HSC']=(df_r['g_SDSS']+df_r['g_HSC'])/2
df_r

df_r.to_csv('./DATA/redshift_bin_org.csv')

# 删掉一些不必要的参数

col=['SDSS_NAME','RA','DEC','hsc_id','ra','dec']
df_r=df_r.drop(col,axis=1)

df_r

df_r.to_csv('./DATA/redshift_bin.csv')

df_s=pd.read_csv('./DATA/redshift_bin.csv',index_col=[0,1])#读取的时候一定要读取两个头哦
