==> AGN_ver1.0/0.4.mean&median&bootstrapping.ipynb <==
# 0

#其他的库
import numpy as np
import pandas as pd
import scipy as sp 
from sklearn.utils import resample
#封装进度条的函数
from tqdm import tqdm   

#导入自己的函数
import mydef

# 1.0

df=pd.read_csv('./DATA/redshift_bin.csv',index_col=[0,1])
df

# mean&median

df_mean_0=pd.DataFrame()
for i in range(52):
    df_mean_0[i]=df.loc[i].mean()
df_mean = pd.DataFrame(df_mean_0.values.T, index=df_mean_0.columns, columns=df_mean_0.index) 
#SDSS_NAME的数据是string所以数据丢失了，不过没有关系
#ra,dec,RA,DEC,HSCid这些数据也是没有意义的
#对于BH的三个对数数据，直接平均也是没有意义的
df_mean.to_csv('./DATA/redshift_bin_mean.csv')

df_median_0=pd.DataFrame()
for i in range(52):
    df_median_0[i]=df.loc[i].median()
df_median = pd.DataFrame(df_median_0.values.T, index=df_median_0.columns, columns=df_median_0.index) 
#SDSS_NAME的数据是string所以数据丢失了，不过没有关系
#ra,dec,RA,DEC,HSCid这些数据也是没有意义的
df_median.to_csv('./DATA/redshift_bin_median.csv')

# mean bootstrapping

#先将红移bin的数据提取出来.#注释一下，df.loc[i]是第i个红移bin
df_list=[]
for i in range(52):
    df_list.append(df.loc[i])

def My_Bs_Mean(list,n_samples=100,n_loop=1000):


    #这里是样本抽一百次的，抽样重复一千次的
    '''
    对一个样本list操作,bootstrapping过程中，每次抽样抽n_samples个，一共抽n_loop次
    返回bootstrapping的平均值的分布列表（未排序），是一个n_loop长度的list
    '''
    i=n_loop
    n=n_samples
    bs_mean=[];
    for i in range(i):##!!循环次数在这里
        bootstrap_samples = resample(list,n_samples=n ,replace=1)#抽样个数再n——sample里
        bs_mean.append(np.mean(bootstrap_samples))
    return bs_mean


df_sigma=pd.DataFrame(columns=df.columns+'_err')

for i in tqdm(range(52)):
    t_list=[]
    for j in df.columns:
        Sigma= lambda x : np.sqrt(np.var(My_Bs_Mean( x ),ddof=1)) 
        t_list.append(Sigma(df_list[i][j]))
    t_series=pd.Series(t_list,index=df_sigma.columns)   
    df_sigma=df_sigma.append(t_series,ignore_index=True)

df_sigma.to_csv('./DATA/redshift_bin_meanerr.csv')

# median bootstrapping


def My_Bs_median(list,n_samples=100,n_loop=1000):


    #这里是样本抽一百次的，抽样重复一千次的
    '''
    对一个样本list操作,bootstrapping过程中，每次抽样抽n_samples个，一共抽n_loop次
    返回bootstrapping的中位值的分布列表（未排序），是一个n_loop长度的list
    '''
    i=n_loop
    n=n_samples
    bs_median=[];
    for i in range(i):##!!循环次数在这里
        bootstrap_samples = resample(list,n_samples=n ,replace=1)#抽样个数再n——sample里
        bs_median.append(np.median(bootstrap_samples))
    return bs_median


df_sigma_median=pd.DataFrame(columns=df.columns+'_err')

for i in tqdm(range(52)):
    t_list=[]
    for j in df.columns:
        sigma_median= lambda x : np.sqrt(np.var(My_Bs_median( x ),ddof=1)) 
        t_list.append(sigma_median(df_list[i][j]))
    t_series=pd.Series(t_list,index=df_sigma_median.columns)   
    df_sigma_median=df_sigma_median.append(t_series,ignore_index=True)

df_sigma_median.to_csv('./DATA/redshift_bin_medianerr.csv')
