==> AGN_ver1.0/2.0.bootstrapping_in_all_bins.ipynb <==
# 0.加载

#其他的库
import numpy as np
import pandas as pd
import scipy as sp 
from sklearn.utils import resample
#封装进度条的函数
from tqdm import tqdm   

#导入自己的函数

# 1.0. 打开文件

df=pd.read_csv('./DATA/redshift_bin.csv',index_col=[0,1])

df

# 1.1.定义函数

def My_Bs_Mean(list,n_samples=100,n_loop=1000):


    #这里是样本抽一百次的，抽样重复一千次的
    '''
    对一个样本list操作,bootstrapping过程中，每次抽样抽n_samples个，一共抽n_loop次
    返回bootstrapping的平均值的分布列表（未排序），是一个n_loop长度的list
    '''
    i=n_loop
    n=n_samples
    bs_mean=[];
    for i in range(i):##!!循环次数在这里
        bootstrap_samples = resample(list,n_samples=n ,replace=1)#抽样个数再n——sample里
        bs_mean.append(np.mean(bootstrap_samples))
    return bs_mean
#####################################################################################
def My_Bs_Median(list,n_samples=100,n_loop=1000):


    #这里是样本抽一百次的，抽样重复一千次的
    '''
    对一个样本list操作,bootstrapping过程中，每次抽样抽n_samples个，一共抽n_loop次
    返回bootstrapping的中位值的分布列表（未排序），是一个n_loop长度的list
    '''
    i=n_loop
    n=n_samples
    bs_median=[];
    for i in range(i):##!!循环次数在这里
        bootstrap_samples = resample(list,n_samples=n ,replace=1)#抽样个数再n——sample里
        bs_median.append(np.median(bootstrap_samples))
    return bs_median

# 1.2.关于分bin继续定义

df_list=[]
for i in range(52):
    df_list.append(df.loc[i])

def My_Bs_Mean_Res_Bin(res):
    
    #第一步，从原始数据中提取分bin的dataframe
    for i in range(52):
        df_list[i].sort_values(by=res,ascending=True,inplace=True)
        df_list[i].reset_index(drop=True,inplace=True)
        df_list[i]['%s_bin_index'% res]=pd.cut(df_list[i].index,bins=np.arange(0,len(df_list[i])+1,len(df_list[i])/5),right=False,labels=range(5))#RIGHT=FALSE表示左闭右开
        df_list[i]=df_list[i].set_index('%s_bin_index'% res)
        
    #第二部，对每一个分bin进行bootstrapping运算，聚合成表（分5个bin所以5个dataframe）
    df_sigma_list=[]#5个df组成的列
    df_mean_list=[]
    df_all_list=[]
    for j in range(5):   
        df_mean_list .append(pd.DataFrame(columns=df.columns))
        df_sigma_list.append(pd.DataFrame(columns=df.columns+'_err'))
        
        for i in tqdm(range(52)):
            
            df_t=df_list[i].loc[j]
                             
            #mean处理
            df_mean_list[j]=df_mean_list[j].append(df_t.mean(),ignore_index=True)
                             
            #bootstrapping处理                     
            t_list=[]
            for k in df_t.columns:
                Sigma= lambda x : np.sqrt(np.var(My_Bs_Mean( x ),ddof=1)) 
                t_list.append(Sigma(df_t[k]))
            t_series=pd.Series(t_list,index=df_sigma_list[j].columns)
            df_sigma_list[j]=df_sigma_list[j].append(t_series,ignore_index=True) 
            
        #所有的都append完之后拼接起来
        df_all_list.append(pd.concat([df_mean_list[j],df_sigma_list[j]],axis=1))
            
    #第三步，聚合成一个表以存储
    df_all=pd.DataFrame(columns=df_all_list[0].columns)
    for j in range(5):
        df_all_list[j]['percent_index'] = j
        
        df_all=df_all.append(df_all_list[j])
    df_all=df_all.set_index(['percent_index',df_all.index])
    return df_all                        


#哎呀妈呀累死我了这列表操作写的,经过我精妙绝伦的编程技巧重写的程序莫名其妙的计算时间成功变成了五倍，淦

# 1.3.执行运算

res='g_SDSS'
d=My_Bs_Mean_Res_Bin(res)
d.to_csv('./DATA/g_SDSS_bin_in_redshift_bin_mean&err.csv')

把剩下几个都算了吧

names=['g_HSC','MBH','EDD_RATIO']
for res in names:    
    d=My_Bs_Mean_Res_Bin(res)
    d.to_csv('./DATA/%s_bin_in_redshift_bin_mean&err.csv'%res)

names=['LBOL','half_g_SDSS+HSC']
for res in names:    
    d=My_Bs_Mean_Res_Bin(res)
    d.to_csv('./DATA/%s_bin_in_redshift_bin_mean&err.csv'%res)

df
