==> AGN_ver1.0/3.2.select_data_change the most.ipynb <==
# matpltlib 库以及参数
import matplotlib
import matplotlib.pyplot as plt
from matplotlib.colors import LogNorm
matplotlib.rcParams.update({'font.size': 20})
matplotlib.rcParams.update({'axes.labelsize': 'large'})
matplotlib.rcParams.update({'xtick.labelsize': 'large'})
matplotlib.rcParams.update({'ytick.labelsize': 'large'})
matplotlib.rcParams.update({'legend.fontsize': 'large'})
matplotlib.rcParams.update({'errorbar.capsize': 3})

#其他的库
import numpy as np
import pandas as pd
import scipy as sp 
from scipy.optimize import leastsq
#封装进度条的函数
from tqdm import tqdm  

COLORLIST1=['#0404B4','#0000ff','#6E6E6E','#ff0000','#B40404']
COLORLIST2=['#0000ff','#0404B4','#000000','#B40404','#ff0000']
COLORLIST3=['#0052ad','#0074f5','#68aefc','#828282','#000000']
COLORLIST4=['#ff0000','#ff7474','#a9a9a9','#848484','#000000']
MARKERLIST=['o','v','x','^','s']


df=pd.read_csv('./DATA/redshift_bin_org.csv',index_col=[0,1])
df_list=[]
for i in range(52):
    ##切开，赋值
    df_list.append(df.loc[i])
    #重排序
    df_list[i].sort_values(by=['g_SDSS-HSC'],ascending=True,inplace=True)
    #重命名索引
    df_list[i].reset_index(drop=True,inplace=True)


n=5
df_list_first=[]
df_list_last=[]
for i in range(52):
    df_list_first.append(df_list[i].iloc[0:n])
    df_list_last.append(df_list[i].iloc[100-n:100])


df_list_last[1]


first=pd.DataFrame(columns=df_list_first[0].columns)
for i in range(52):
    first=first.append(df_list_first[i])
    first.reset_index(drop=True,inplace=True)

first
